% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{fs_search}
\alias{fs_search}
\title{FactSet Filesystem Search}
\usage{
fs_search(con, dir, type = c("pa", "spar", "dir", "accounts"))
}
\arguments{
\item{con}{A FactSet connection object.}

\item{dir}{A directory in the FactSet filesystem associated with your serial number.}

\item{type}{The type of item you want to search for. Right now the options include accounts, directories, PA documents, or SPAR documents.}
}
\value{
A tibble containing the accounts, documents, or directories within the specified directory.
}
\description{
This function makes it easy to search your FactSet filesystem. It returns a tibble with the accounts, directories, or documents in the specified directory.
}
\examples{
fs_pa_docs <- fs_search(con, dir, type = "pa")
}
\seealso{
See \link{fs_auth} for authentication instructions.
}
